Separa service type en su propio provider
