+++ lib/provider/pet_service.dart
@@ -1,7 +1,8 @@
-import 'package:dio/dio.dart';
-import 'package:firulapp/constants/endpoints.dart';
-import 'package:firulapp/provider/user.dart';
 import 'package:flutter/foundation.dart';
+
+import 'package:dio/dio.dart';
+import '../constants/endpoints.dart';
+import './user.dart';
 
 class PetServiceItem {
   String id;
@@ -127,46 +128,7 @@ class PetService with ChangeNotifier {
     );
   }
 
-  PetServiceItem getLocalPetServiceById(String serviceId) {}
-}
-
-class CategoryItem {
-  final String id;
-  final String title;
-  final String icon;
-
-  const CategoryItem({this.id, this.title, this.icon});
-
-  static const DUMMY_CATEGORIES = const [
-    CategoryItem(
-      id: '1',
-      title: 'Baño',
-      icon: 'assets/icons/dog-shower.svg',
-    ),
-    CategoryItem(
-      id: '2',
-      title: 'Paseo',
-      icon: 'assets/icons/dog-walking.svg',
-    ),
-    CategoryItem(
-      id: '3',
-      title: 'Tienda',
-      icon: 'assets/icons/pet-shop.svg',
-    ),
-    CategoryItem(
-      id: '4',
-      title: 'Entrenamiento',
-      icon: 'assets/icons/dog-train.svg',
-    ),
-    CategoryItem(
-      id: '5',
-      title: 'Veterinaria',
-      icon: 'assets/icons/vet.svg',
-    ),
-    CategoryItem(
-      id: '0',
-      title: 'Mis Servicios',
-      icon: 'assets/icons/businessman.svg',
-    ),
-  ];
+  PetServiceItem getLocalPetServiceById(String serviceId) {
+    return _items.firstWhere((serv) => serv.id == serviceId);
+  }
 }
