+++ lib/src/pet_services/pet_service_form.dart
@@ -5,7 +5,8 @@ import 'package:firulapp/components/dialogs.dart';
 import 'package:firulapp/provider/species.dart';
 import 'package:firulapp/provider/user.dart';
 import 'package:firulapp/components/input_text.dart';
-import 'package:firulapp/provider/pet_service.dart';
+import '../../provider/pet_service.dart';
+import '../../provider/service_type.dart';
 import 'package:firulapp/src/mixins/validator_mixins.dart';
 import '../../components/default_button.dart';
 import '../../size_config.dart';
